# Optimize iterstraight