This took me a while, as I had some false starts. Reading a 800x800 image from memory was taking ~130ms on my laptop, the main offenders being:

iterstraight - 50ms
cpngfilters.undo_filter_paeth - 37ms
{built-in method decompress} - 16ms
...

iterstraight spends most of its time extending the a array to then pass a slice of it. Very often (depends on the IDAT chunks) it can just pass a slice the "some" array. I take advantage of this to create a fast-path for this case (the while loop).

This takes iterstraight from 50ms down to 2ms, putting undo_filter_\* back as the top offenders.

I found a way to make undo_filter_\* faster, specifying in Cython that the elements over which it iterates are contiguous in memory. This takes undo_filter_paeth down to 25ms.

In total this pull request saves ~60ms, making pypng almost twice as fast as it was before (in total ~70ms to load a 800x800 image from memory).